refactor(install): preserve service module budgets
