+++ src/apm_cli/install/services.py
@@ -23,6 +23,10 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
+from .deployed_paths import deployed_path_entry as _deployed_path_entry
+from .deployed_paths import skill_bundle_file_entries as _skill_bundle_file_entries
+from .target_filter import filter_targets_for_dependency
+
 if TYPE_CHECKING:
     from ..core.command_logger import InstallLogger
     from ..core.scope import InstallScope
@@ -59,72 +63,6 @@ class IntegratorBundle:
     # keep working. Production sites (template.py, integrate_local_content,
     # drift.py) pass a real CanvasIntegrator; when None the loop skips canvas.
     canvas: BaseIntegrator | None = None
-
-
-def _deployed_path_entry(
-    target_path: Path,
-    project_root: Path,
-    targets: Any,
-) -> str:
-    """Return the lockfile-safe path string for a deployed file."""
-
-    def _try_dynamic_root(tgts, *, strict: bool = False) -> str | None:
-        for _t in tgts:
-            if _t.resolved_deploy_root is None:
-                continue
-            if not strict:
-                try:
-                    target_path.relative_to(_t.resolved_deploy_root)
-                except ValueError:
-                    continue
-            if _t.name == "copilot-app":
-                from apm_cli.integration.copilot_app_db import to_lockfile_uri
-
-                return to_lockfile_uri(target_path.name)
-            from apm_cli.integration.copilot_cowork_paths import to_lockfile_path
-
-            return to_lockfile_path(target_path, _t.resolved_deploy_root)
-        return None
-
-    if targets:
-        result = _try_dynamic_root(targets)
-        if result is not None:
-            return result
-    try:
-        return target_path.relative_to(project_root).as_posix()
-    except ValueError:
-        # Fallback: let to_lockfile_path run its own security
-        # validation (PathTraversalError) without pre-filtering.
-        if targets:
-            result = _try_dynamic_root(targets, strict=True)
-            if result is not None:
-                return result
-        raise RuntimeError(  # noqa: B904
-            f"Cannot translate {target_path!r} to a lockfile path: "
-            f"path is outside the project tree and no dynamic-root "
-            f"target matched. This is a bug -- please report it."
-        )
-
-
-def _skill_bundle_file_entries(
-    skill_dir: Path,
-    project_root: Path,
-    targets: Any,
-) -> list[str]:
-    """Expand a deployed skill directory into per-file lockfile entries."""
-    try:
-        if not (skill_dir.is_dir() and not skill_dir.is_symlink()):
-            return []
-    except OSError:
-        return []
-    entries: list[str] = []
-    for bundle_file in sorted(skill_dir.rglob("*")):
-        try:
-            if bundle_file.is_file() and not bundle_file.is_symlink():
-                entries.append(_deployed_path_entry(bundle_file, project_root, targets))
-        except OSError:
-            continue
-    return entries
 
 
 def _log_hook_display_payloads(
@@ -248,26 +186,6 @@ def _log_canvas_skip(package_name: str, package_info: Any, logger: InstallLogger
         )
 
 
-def _filter_targets_for_dependency(
-    targets: Any,
-    dep_target_subset: list[str] | None,
-    diagnostics: DiagnosticCollector,
-    package_name: str,
-) -> tuple[Any, set[str], bool]:
-    """Apply the consumer-manifest dependency target filter."""
-    allowed_dep_targets = builtins.set(dep_target_subset or [])
-    if not dep_target_subset:
-        return targets, allowed_dep_targets, False
-
-    filtered_targets = [target for target in targets if target.name in allowed_dep_targets]
-    if not filtered_targets:
-        diagnostics.warn(
-            "per-dependency targets do not overlap active install targets; skipping",
-            package=package_name,
-        )
-    return filtered_targets, allowed_dep_targets, True
-
-
 def integrate_package_primitives(  # noqa: PLR0913
     package_info: Any,
     project_root: Path,
@@ -331,7 +249,7 @@ def integrate_package_primitives(  # noqa: PLR0913
 
     # SECURITY: dep_target_subset comes from CONSUMER manifest only.
     # Package-side targets are advisory metadata; never a routing input.
-    targets, allowed_dep_targets, dep_targets_active = _filter_targets_for_dependency(
+    targets, allowed_dep_targets, dep_targets_active = filter_targets_for_dependency(
         targets,
         dep_target_subset,
         diagnostics,
